+++ packages/form-render/src/useForm.js
@@ -110,10 +110,6 @@ export const useForm = props => {
 
   // 这里导致第二次的渲染
   useEffect(() => {
-    if (clickSubmit.current) {
-      clickSubmit.current = false;
-      return;
-    }
     validateAll({
       formData: _data.current,
       schema: schemaRef.current,
@@ -191,7 +187,6 @@ export const useForm = props => {
   };
 
   const submit = () => {
-    clickSubmit.current = true;
     setState({ isValidating: true, allTouched: true, isSubmitting: false });
     //  https://formik.org/docs/guides/form-submission
     // TODO: 更多的处理，注意处理的时候一定要是copy一份formData，否则submitData会和表单操作实时同步的。。而不是submit再变动了
@@ -245,6 +240,7 @@ export const useForm = props => {
   const resetFields = () => {
     setState({
       formData: {},
+      submitData: {},
       errors: [],
       touchedKeys: [],
       allTouched: false,
@@ -273,15 +269,6 @@ export const useForm = props => {
       outsideValidating: false,
     });
 
-  const destroyForm = () =>
-    setState({
-      formData: {},
-      submitData: {},
-      errorFields: [],
-      allTouched: false,
-      touchedKeys: [],
-    });
-
   const form = {
     // state
     formData: _data.current,
@@ -308,7 +295,6 @@ export const useForm = props => {
     isEditing,
     setEditing,
     syncStuff,
-    destroyForm,
   };
   return form;
 };