+++ tools/schema-generator/src/FRWrapper.js
@@ -18,8 +18,7 @@ import { Ctx, StoreCtx } from './context';
 // import SCHEMA from './json/basic.json';
 import FR from './FR';
 import PreviewFR from './PreviewFR';
-import { Modal, Input, message } from 'antd';
-import { Button } from 'antd';
+import { Modal, Input, Button, message } from 'antd';
 
 const { TextArea } = Input;
 
@@ -58,20 +57,20 @@ function Wrapper(
   const flatten = flattenSchema(_schema);
   const flattenWithData = dataToFlatten(flatten, formData);
 
-  const onFlattenChange = (newFlatten, changeSource = 'schema') => {
+  const onFlattenChange = (newFlatten) => {
     const newSchema = idToSchema(newFlatten);
     const newData = flattenToData(newFlatten);
     // 判断只有schema变化时才调用，一般需求的用户不需要
-    if (changeSource === 'schema' && onSchemaChange) {
+    if (onSchemaChange) {
       onSchemaChange(newSchema);
     }
     // schema 变化大都会触发 data 变化
     onChange(newData);
   };
 
-  const onItemChange = (key, value, changeSource) => {
+  const onItemChange = (key, value) => {
     flattenWithData[key] = value;
-    onFlattenChange(flattenWithData, changeSource);
+    onFlattenChange(flattenWithData);
   };
 
   const toggleModal = () => setState({ showModal: !local.showModal });